requirements.txt
flask
requests
beautifulsoup4
googlesearch-python
peewee
newspaper3k
sentence-transformers
nltk
openai